FE: Move side nav components around
