+++ frontend/src/lib/figma/navigation/side-nav/FilterLabelsClosed.svelte
